rename scripts without orchestration
