+++ script_without_orchestration/credentials.py
@@ -1,0 +1,10 @@
+"""
+Credentials for sending inspiring quotes with image to MS Teams channel
+"""
+PEXEL_API_KEY = "Your Pexel API Key"    # Enter your Pexel API key
+
+WEBHOOK_TEAMS = "Your Webhook"  # Enter your MS Teams Webhook Url
+
+QUOTE_URL = "https://zenquotes.io/api/random"   # Enter zenquotes.io API Url
+
+MESSAGE_SENDER_NAME = "Name Surname"  # Enter in format <Name Surname>
