+++ script_without_orchestration/custom_errors.py
@@ -1,0 +1,43 @@
+"""
+Script with custom Exceptions and Errors for the Mental State Improvement Project
+"""
+
+
+class BaseMSIProjectException(Exception):
+    pass
+
+
+class ImageException(BaseMSIProjectException):
+    pass
+
+
+class APINotAvailable(ImageException):
+    pass
+
+
+class PexelAPIKeyError(ImageException):
+    pass
+
+
+class JsonError(ImageException):
+    pass
+
+
+class QuoteException(BaseMSIProjectException):
+    pass
+
+
+class NoResponseFromQuoteUrl(QuoteException):
+    pass
+
+
+class JSONExtractError(QuoteException):
+    pass
+
+
+class MessageException(BaseMSIProjectException):
+    pass
+
+
+class WebhookUrlError(MessageException):
+    pass
